+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## Added
+
+- `EveSolarSystem.nearest_celestial` can now also returns stars
+
+## Changed
+
+- Switch from evemicros to evesdeapi for calculating nearest celestials
+- Drop support for Python 3.7
+
 ## [0.18.0] - 2022-09-19
 
 ## Added